refactor: Move TIntMaskVerifier operations to dedicated TIntMaskVerifierAnalyzer class
